LCL: TLMMouse.Keys:PtrInt becomes TLMMouse.Keys: PtrUInt to avoid range check errors. Qt, Gtk2: added various type casts to avoid range check errors. Patches by Juha Manninen. fixes #17531
git-svn-id: http://svn.freepascal.org/svn/lazarus/trunk@27786 4005530d-fff6-0310-9dd1-cebe43e6787f

+++ lcl/interfaces/gtk2/gtk2cellrenderer.pas
@@ -127,14 +127,14 @@ begin
   if ItemIndex < 0 then
     ItemIndex := 0;
 
-  MeasureItemStruct.itemID := ItemIndex;
-  MeasureItemStruct.itemWidth := width^;
-  MeasureItemStruct.itemHeight := height^;
+  MeasureItemStruct.itemID := UINT(ItemIndex);
+  MeasureItemStruct.itemWidth := UINT(width^);
+  MeasureItemStruct.itemHeight := UINT(height^);
   Msg.Msg := LM_MEASUREITEM;
   Msg.MeasureItemStruct := @MeasureItemStruct;
   DeliverMessage(AWinControl, Msg);
-  width^ := MeasureItemStruct.itemWidth;
-  height^ := MeasureItemStruct.itemHeight;
+  width^ := gint(MeasureItemStruct.itemWidth);
+  height^ := gint(MeasureItemStruct.itemHeight);
 end;
 
 function GtkCellRendererStateToListViewDrawState(CellState: TGtkCellRendererState): TCustomDrawState;
@@ -271,7 +271,7 @@ begin
   try
     FillChar(Msg.DrawListItemStruct^,SizeOf(TDrawListItemStruct),0);
     with Msg.DrawListItemStruct^ do begin
-      ItemID:=ItemIndex;
+      ItemID:=UINT(ItemIndex);
       Area:=AreaRect;
       //DebugLn(['LCLIntfCellRenderer_Render Widget=',GetWidgetDebugReport(Widget^.parent),' Area=',dbgs(Area)]);
       DCWidget:=Widget;
